Move API code into dedicated file
